Minor refactoring and spell error
