+++ client/src/actions/index.jsx
@@ -1,4 +1,4 @@
-import { CHANGE_INFO } from './actionsTypes';
+import { CHANGE_INFO } from './actionTypes';
 
 // Action Creators for Game Info page: 
 export function newInfoSection(pageName = 'GETTING STARTED') {
