+++ client/src/actions/setGameState.js
@@ -3,6 +3,5 @@ import { CHANGE_GAME_STATE } from './actionTypes';
 export default function setGameState(){
   return{
     type: CHANGE_GAME_STATE
-    }
   }
 }
